tools/requirements.txt
torch==2.0.1
torchvision==0.15.2
faster-coco-eval>=1.6.6
PyYAML
tensorboard
scipy
pycocotools
onnx==1.14.0
onnxruntime-gpu
onnxscript
onnxsim==0.4.33